requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.39.0
python-chess==1.999
requests==2.32.3
pydantic==2.9.2
